==> notebooks/03_EDA_Ember_Full_Release_Data.ipynb <==
#### Import & Path Setups

import sys
from pathlib import Path

import pandas as pd

# visualization
import plotly.express as px
import plotly.express as px
import plotly.graph_objects as go

PROJECT_ROOT = Path.cwd().parents[0]
SRC_DIR = PROJECT_ROOT / "src"
if str(SRC_DIR) not in sys.path:
    sys.path.insert(0, str(SRC_DIR))

from my_project.paths import get_paths

paths = get_paths(PROJECT_ROOT)
DATA_DIR = paths['DATA_DIR']
RAW_DATA_DIR = paths['RAW_DATA_DIR']
PROCESSED_DATA_DIR = paths['PROCESSED_DATA_DIR']
LOGS_DIR = paths['LOGS_DIR']

from utils.plots import save_static_img
from data.get_data import download_from_url

#### Download Dataset

file_url = "https://storage.googleapis.com/emb-prod-bkt-publicdata/public-downloads/yearly_full_release_long_format.csv"
yearly_full_release = download_from_url(file_url, RAW_DATA_DIR)
df = pd.read_csv(yearly_full_release)
df.sample(3)

#### Transformations

df.info()

df.sample(3)

# data cleaning for column names in df
col_name_clean = df.columns.str.lower().str.replace(" ", "_")
df.columns = col_name_clean

# rename cols to be more descriptive
df = df.rename(columns={
    "variable": "energy_type",
})

# remove unused columns
df = df.drop(columns=["ember_region", "eu", "oecd", "g20", "g7", "asean", "yoy_absolute_change", "yoy_%_change"])

# drop % rows from unit column
df = df[df["unit"] != "%"]

# remove unwanted subcategories 
# removing "Total" since it is an aggregate of generation and emissions
# removing "Aggregate Fuel": we'll use unaggregated data for more granularity in our analysis.
df = df[~df["subcategory"].isin(["Total", "Aggregate fuel"])]

# split dataset based on area_type
df_country = df[df["area_type"] == "Country or economy"]
df_region = df[df["area_type"] == "Region"]
# drop the area_type column as it's no longer needed
df_country = df_country.drop(columns=["area_type"])
df_region = df_region.drop(columns=["area_type"])

##### Country/Economy Data

df_country.sample(3)
